pyproject.toml
[project]
name = "rag-langchain-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-google-genai>=4.2.0",
    "langchain-pymupdf4llm>=0.5.0",
    "langchain[google-genai]>=1.2.9",
    "loguru>=0.7.3",
    "markdown>=3.10.1",
    "pypdf>=6.6.2",
    "pytesseract>=0.3.13",
    "streamlit>=1.54.0",
    "unstructured>=0.18.31",
]
